Use `projects_tag` as full URL
There was a discrepancy between the topics regex (which expected
`projects_tag` to be a full Github URL) and the line concatenating it to
the end of the Github URL prefix (which expected it to only be the tag).

I resolved the discrepancy in favor of the full URL in the
`projects_tag` field (in part to be more explicit as well as to support
multiple Git hosts beyond Github someday), but missed updating this
variable.