Add `last_pushed_within` field to project index
For multiple types of our users (CfA Staff, Brigade Leader, Project
Leaders), being able to tell which projects are still active is a
crucial aspect of the index.

In #26 we discuss using a bucketed approach so as to not create
unnecessary noise by committing the timestamp for every update. This
commit implements a coarse timestamp: for projects updated within the
last week, month, year, or over a year ago.